+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### To Release v0.4.0
 
+* c+; comment selected text(or that line if no selected text) if uncommented, uncomment if commented
 * o+backspace delete previous sexp
 * c+backspace delete whole line from current position